drivers: msm: undervolt SVS/NOM rails
Dropping the voltage on the SVS rail to 1200000 uV (micro-Volts) or 1.2 Volts (V) from 1800000 uV (micro-Volts) or 1.8 Volts (V). SVS rail doesn't really need that much of power in order to work and power anything on it. Dropping the voltage on the NOM rail to 1450000 uV (micro-Volts) or 1.45 Volts (V) from 1800000 uV (micro-Volts) or 1.8 Volts (V). Nor SVS nor NOM rails really need that much of power in order to work and power anything on them. Note that a Micro-optimization inside the variable frequencies (CPU,GPU) is needed in order to gain some better power efficiency.

Signed-off-by: DragO0nFly <greekgod97ak47@gmail.com>