drivers: msm: undervolt PLL
A phase-locked loop (PLL) is a control system that generates an output signal whose phase is related to the phase of an input signal, a PLL is also used to synthesize new frequencies which are a multiple of a reference frequency (XO), with the same stability as the reference frequency. There are several different types; the simplest is an electronic circuit consisting of a variable frequency oscillator and a phase detector in a feedback loop. The oscillator generates a periodic signal, and the phase detector compares the phase of that signal with the phase of the input periodic signal, adjusting the oscillator to keep the phases matched. a53sspll is an Analog or linear PLL (APLL) which uses a voltage-controlled oscillator (VCO) with a passive Loop Filter. The VCO is locked to a Frequency modulation (FM). Since the PLL is locked to a FM singal, the VCO tracks the instantaneous frequency of the input signal. In most scenarios the used VCO in the PLL is highly linear which let's us create linear frequency modulators. As of in this case, the PLL is about to supply 900000000 nano-volts (nV) or 0.9 Volts (V) on the SVS rail instead of 1000000000 nano-volts (nV) or 1 Volt (V). On the other hand, the PLL will also supply 1365000000 nano-volts (nV) or 1.365 Volts (V) on the TUR (TURBO) rail from 1900000000 nano-volts (nV) or 1.9 Volts (V). Note, lower voltages might bring instability and various CPU frequency related issues. The PLL might not have enough power, in order to execute certain frequencies.

Signed-off-by: DragO0nFly <greekgod97ak47@gmail.com>